--- .env.example
@@ -1,7 +1,0 @@
-DB_HOST=localhost
-DB_USER=labber
-DB_PASS=labber
-DB_NAME=midterm
-# Uncomment and set to true for Heroku
-# DB_SSL=true if heroku
-DB_PORT=5432
