--- docker-compose.yml
@@ -1,34 +1,0 @@
-version: "3.4"
-
-services:
-  quiz-app:
-    image: quiz-app
-    build:
-      context: .
-      dockerfile: ./Dockerfile
-    environment:
-      NODE_ENV: production
-      DB_HOST: postgres
-      DB_USER: username
-      DB_PASS: password
-      DB_NAME: database
-      DB_PORT: 5432
-    ports:
-      - 8080:8080
-  postgres:
-    image: postgres
-    restart: always
-    ports:
-      - 5432:5432
-    environment:
-      POSTGRES_PASSWORD: password
-      POSTGRES_USER: username
-      POSTGRES_DB: database
-    tmpfs:
-      - /var/lib/postgresql/data
-    # volumes:
-    #   - db_volume:/var/lib/postgresql/data
-volumes:
-  db_volume:
-
-    #psql postgresql://username:password@postgres:5432/database
