+++ services/backend/.env.example
@@ -1,0 +1,7 @@
+POSTGRES_HOST=localhost
+POSTGRES_USER=labber
+POSTGRES_PASSWORD=labber
+POSTGRES_DB=midterm
+# Uncomment and set to true for Heroku
+# POSTGRES_SSL=true if heroku
+POSTGRES_PORT=5432
