+++ services/backend/docker-compose.yml
@@ -1,0 +1,34 @@
+version: "3.4"
+
+services:
+  quiz-app:
+    image: quiz-app
+    build:
+      context: .
+      dockerfile: ./Dockerfile
+    environment:
+      NODE_ENV: production
+      POSTGRES_HOST: postgres
+      POSTGRES_USER: username
+      POSTGRES_PASSWORD: password
+      POSTGRES_DB: database
+      POSTGRES_PORT: 5432
+    ports:
+      - 8080:8080
+  postgres:
+    image: postgres
+    restart: always
+    ports:
+      - 5432:5432
+    environment:
+      POSTGRES_PASSWORD: password
+      POSTGRES_USER: username
+      POSTGRES_DB: database
+    tmpfs:
+      - /var/lib/postgresql/data
+    # volumes:
+    #   - POSTGRES_volume:/var/lib/postgresql/data
+volumes:
+  POSTGRES_volume:
+
+    #psql postgresql://username:password@postgres:5432/database
