+++ services/backend/package.json
@@ -5,24 +5,18 @@
   "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js",
-    "local": "./node_modules/.bin/nodemon -L --ignore public/",
-    "db:reset": "node bin/resetdb.js"
+    "start": "node server.js"
   },
   "author": "Lighthouse Labs",
   "license": "ISC",
   "dependencies": {
     "body-parser": "^1.19.0",
-    "chalk": "^2.4.2",
-    "dotenv": "^2.0.0",
-    "ejs": "^2.6.2",
+    "dotenv": "^8.2.0",
+    "ejs": "^3.1.6",
     "express": "^4.17.1",
-    "morgan": "^1.9.1",
+    "morgan": "^1.10.0",
     "node-sass-middleware": "^0.11.0",
-    "pg": "^6.4.2",
+    "pg": "^8.5.1",
     "pg-native": "^3.0.0"
-  },
-  "devDependencies": {
-    "nodemon": "^1.19.1"
   }
 }
