+++ services/backend/server.js
@@ -1,5 +1,7 @@
-// load .env data into process.env
-// require('dotenv').config();
+try {
+  // load .env data into process.env
+  require('dotenv').config();
+} catch (e) { }
 
 // Web server config
 const PORT = process.env.PORT || 8080;
@@ -11,7 +13,7 @@ const app = express();
 const morgan = require('morgan');
 
 // PG database client/connection setup
-const { Pool } = require('pg');
+const { Pool } = require('pg').native;
 const dbParams = require('./lib/db.js');
 const db = new Pool(dbParams);
 db.connect();
